{"schema":"decodex/commit/1","summary":"Modularize intake proposal research and site style boundaries","authority":"manual"}
