+++ site/src/styles/appcast/panel.css
@@ -1,0 +1,48 @@
+.appcast-download__panel {
+  position: absolute;
+  right: 0;
+  top: calc(100% + 0.54rem);
+  bottom: auto;
+  z-index: 120;
+  display: grid;
+  gap: 0.42rem;
+  width: min(13.6rem, calc(100vw - 2rem));
+  min-width: min(13rem, calc(100vw - 2rem));
+  padding: 0.62rem;
+  border: 1px solid var(--appcast-menu-border);
+  border-radius: 0.65rem;
+  background: var(--appcast-menu-bg);
+  box-shadow: var(--appcast-menu-shadow);
+  color: var(--appcast-menu-ink);
+  text-align: left;
+  clip-path: inset(0 0 7% 0 round 0.65rem);
+  filter: blur(2px) saturate(0.92);
+  opacity: 0;
+  overflow: visible;
+  pointer-events: none;
+  transform: translateY(-0.22rem) scale(0.985);
+  transform-origin: top right;
+  transition:
+    clip-path 220ms cubic-bezier(0.2, 0.8, 0.2, 1),
+    filter 220ms ease,
+    opacity 160ms ease,
+    transform 220ms cubic-bezier(0.2, 0.8, 0.2, 1);
+}
+
+.appcast-download[data-appcast-open="true"] .appcast-download__panel,
+.appcast-download:hover .appcast-download__panel,
+.appcast-download:focus-within .appcast-download__panel {
+  clip-path: inset(0 0 0 0 round 0.65rem);
+  filter: blur(0) saturate(1);
+  opacity: 1;
+  pointer-events: auto;
+  transform: translateY(0) scale(1);
+}
+
+.appcast-download__panel-label {
+  color: var(--appcast-menu-muted);
+  font-size: 0.72rem;
+  font-weight: 780;
+  line-height: 1.2;
+  text-transform: uppercase;
+}
