+++ site/src/styles/appcast/versions.css
@@ -1,0 +1,42 @@
+.appcast-download__list {
+  display: grid;
+  gap: 0.12rem;
+  max-height: 10.4rem;
+  margin: 0;
+  padding: 0 0.18rem 0 0;
+  overflow-y: auto;
+  list-style: none;
+  scrollbar-width: thin;
+}
+
+.appcast-download__version {
+  display: flex;
+  align-items: baseline;
+  justify-content: space-between;
+  gap: 0.8rem;
+  min-width: 0;
+  min-height: 2rem;
+  padding: 0.32rem 0.4rem;
+  border-radius: 0.45rem;
+  line-height: 1.2;
+  will-change: transform;
+}
+
+.appcast-download__version:hover,
+.appcast-download__version:focus-visible {
+  background: var(--appcast-menu-hover);
+  color: var(--appcast-menu-accent);
+  transform: translateX(-0.1rem);
+}
+
+.appcast-download__version-title {
+  font-size: 0.9rem;
+  font-weight: 780;
+  white-space: nowrap;
+}
+
+.appcast-download__version-meta,
+.appcast-download__empty {
+  color: var(--appcast-menu-muted);
+  font-size: 0.78rem;
+}
