+++ site/src/styles/theme.css
@@ -1,153 +1,3 @@
-:root,
-:root[data-theme="light"] {
-  color-scheme: light;
-  --font-sans: "Inter", "SF Pro Text", -apple-system, BlinkMacSystemFont, "Segoe UI", sans-serif;
-  --font-display: "Inter", "SF Pro Display", -apple-system, BlinkMacSystemFont, "Segoe UI", sans-serif;
-  --font-mono: "SFMono-Regular", "SF Mono", ui-monospace, Menlo, Monaco, Consolas, monospace;
-  --bg: #f4f6f8;
-  --panel: #ffffff;
-  --panel-soft: #eef2f6;
-  --ink: #111720;
-  --ink-soft: #3b4655;
-  --muted: #6b7584;
-  --muted-soft: #8d96a4;
-  --border: #d5dbe4;
-  --border-strong: #b8c2cf;
-  --accent: #2563eb;
-  --accent-strong: #1748b5;
-  --accent-soft: #e7efff;
-  --lime: #b6f13a;
-  --lime-ink: #203300;
-  --stage-bg: #f8fbff;
-  --stage-ink: #07111e;
-  --stage-soft: #33445d;
-  --stage-muted: #6f7b8f;
-  --stage-border: rgba(12, 22, 38, 0.14);
-  --stage-control-bg: rgba(255, 255, 255, 0.58);
-  --stage-control-border: rgba(12, 22, 38, 0.14);
-  --stage-control-hover: rgba(12, 22, 38, 0.06);
-  --stage-action-hover-bg: rgba(12, 22, 38, 0.075);
-  --stage-action-hover-border: rgba(12, 22, 38, 0.22);
-  --stage-action-hover-ink: var(--stage-ink);
-  --stage-primary-action-hover-bg: var(--lime);
-  --stage-primary-action-hover-ink: var(--lime-ink);
-  --stage-glass-bg: rgba(255, 255, 255, 0.54);
-  --stage-glass-bg-hover: rgba(255, 255, 255, 0.7);
-  --stage-selected-bg: #0d1726;
-  --stage-selected-ink: #f7fbff;
-  --appcast-menu-bg: rgba(255, 255, 255, 0.98);
-  --appcast-menu-ink: var(--ink);
-  --appcast-menu-muted: var(--muted);
-  --appcast-menu-border: rgba(12, 22, 38, 0.14);
-  --appcast-menu-hover: rgba(37, 99, 235, 0.08);
-  --appcast-menu-accent: var(--accent-strong);
-  --appcast-menu-shadow: 0 22px 60px rgba(20, 32, 52, 0.16);
-  --code-bg: #e8edf4;
-  --code-border: #c9d2df;
-  --hero-background:
-    linear-gradient(135deg, #f9fbff, #e8f0ff 58%, #f4fbf5);
-  --hero-veil:
-    linear-gradient(90deg, rgba(248, 251, 255, 0.88) 0%, rgba(248, 251, 255, 0.5) 52%, rgba(248, 251, 255, 0.08) 100%);
-  --hero-grid-line: rgba(12, 22, 38, 0.065);
-}
-
-@media (prefers-color-scheme: dark) {
-  :root:not([data-theme]),
-  :root[data-theme="system"] {
-    color-scheme: dark;
-    --bg: #0d1117;
-    --panel: #141a22;
-    --panel-soft: #10161f;
-    --ink: #edf2ff;
-    --ink-soft: #bac4d5;
-    --muted: #8792a4;
-    --muted-soft: #687489;
-    --border: #27303d;
-    --border-strong: #394556;
-    --accent: #8ab4ff;
-    --accent-strong: #aecbff;
-    --accent-soft: #16243b;
-    --lime: #b6f13a;
-    --lime-ink: #162300;
-    --stage-bg: #0c1118;
-    --stage-ink: #f5f7ff;
-    --stage-soft: #b9c2d4;
-    --stage-muted: #778397;
-    --stage-border: rgba(255, 255, 255, 0.14);
-    --stage-control-bg: rgba(15, 22, 32, 0.72);
-    --stage-control-border: rgba(255, 255, 255, 0.14);
-    --stage-control-hover: rgba(255, 255, 255, 0.08);
-    --stage-action-hover-bg: rgba(182, 241, 58, 0.1);
-    --stage-action-hover-border: rgba(182, 241, 58, 0.48);
-    --stage-action-hover-ink: var(--lime);
-    --stage-primary-action-hover-bg: var(--lime);
-    --stage-primary-action-hover-ink: var(--lime-ink);
-    --stage-glass-bg: rgba(10, 15, 23, 0.42);
-    --stage-glass-bg-hover: rgba(10, 15, 23, 0.58);
-    --stage-selected-bg: #f5f7ff;
-    --stage-selected-ink: #111720;
-    --appcast-menu-bg: rgba(11, 17, 26, 0.98);
-    --appcast-menu-ink: #f5f7ff;
-    --appcast-menu-muted: #778397;
-    --appcast-menu-border: rgba(255, 255, 255, 0.14);
-    --appcast-menu-hover: rgba(255, 255, 255, 0.08);
-    --appcast-menu-accent: var(--lime);
-    --appcast-menu-shadow: 0 22px 60px rgba(0, 0, 0, 0.38);
-    --code-bg: #1e2632;
-    --code-border: #344154;
-    --hero-background:
-      linear-gradient(135deg, #0a0f16, #111923 58%, #071018);
-    --hero-veil:
-      linear-gradient(90deg, rgba(10, 15, 22, 0.76) 0%, rgba(10, 15, 22, 0.38) 58%, rgba(10, 15, 22, 0.08) 100%);
-    --hero-grid-line: rgba(255, 255, 255, 0.035);
-  }
-}
-
-:root[data-theme="dark"] {
-  color-scheme: dark;
-  --bg: #0d1117;
-  --panel: #141a22;
-  --panel-soft: #10161f;
-  --ink: #edf2ff;
-  --ink-soft: #bac4d5;
-  --muted: #8792a4;
-  --muted-soft: #687489;
-  --border: #27303d;
-  --border-strong: #394556;
-  --accent: #8ab4ff;
-  --accent-strong: #aecbff;
-  --accent-soft: #16243b;
-  --lime: #b6f13a;
-  --lime-ink: #162300;
-  --stage-bg: #0c1118;
-  --stage-ink: #f5f7ff;
-  --stage-soft: #b9c2d4;
-  --stage-muted: #778397;
-  --stage-border: rgba(255, 255, 255, 0.14);
-  --stage-control-bg: rgba(15, 22, 32, 0.72);
-  --stage-control-border: rgba(255, 255, 255, 0.14);
-  --stage-control-hover: rgba(255, 255, 255, 0.08);
-  --stage-action-hover-bg: rgba(182, 241, 58, 0.1);
-  --stage-action-hover-border: rgba(182, 241, 58, 0.48);
-  --stage-action-hover-ink: var(--lime);
-  --stage-primary-action-hover-bg: var(--lime);
-  --stage-primary-action-hover-ink: var(--lime-ink);
-  --stage-glass-bg: rgba(10, 15, 23, 0.42);
-  --stage-glass-bg-hover: rgba(10, 15, 23, 0.58);
-  --stage-selected-bg: #f5f7ff;
-  --stage-selected-ink: #111720;
-  --appcast-menu-bg: rgba(11, 17, 26, 0.98);
-  --appcast-menu-ink: #f5f7ff;
-  --appcast-menu-muted: #778397;
-  --appcast-menu-border: rgba(255, 255, 255, 0.14);
-  --appcast-menu-hover: rgba(255, 255, 255, 0.08);
-  --appcast-menu-accent: var(--lime);
-  --appcast-menu-shadow: 0 22px 60px rgba(0, 0, 0, 0.38);
-  --code-bg: #1e2632;
-  --code-border: #344154;
-  --hero-background:
-    linear-gradient(135deg, #0a0f16, #111923 58%, #071018);
-  --hero-veil:
-    linear-gradient(90deg, rgba(10, 15, 22, 0.76) 0%, rgba(10, 15, 22, 0.38) 58%, rgba(10, 15, 22, 0.08) 100%);
-  --hero-grid-line: rgba(255, 255, 255, 0.035);
-}
+@import "./theme/light.css";
+@import "./theme/system-dark.css";
+@import "./theme/dark.css";
