+++ site/src/styles/theme/light.css
@@ -1,0 +1,52 @@
+:root,
+:root[data-theme="light"] {
+  color-scheme: light;
+  --font-sans: "Inter", "SF Pro Text", -apple-system, BlinkMacSystemFont, "Segoe UI", sans-serif;
+  --font-display: "Inter", "SF Pro Display", -apple-system, BlinkMacSystemFont, "Segoe UI", sans-serif;
+  --font-mono: "SFMono-Regular", "SF Mono", ui-monospace, Menlo, Monaco, Consolas, monospace;
+  --bg: #f4f6f8;
+  --panel: #ffffff;
+  --panel-soft: #eef2f6;
+  --ink: #111720;
+  --ink-soft: #3b4655;
+  --muted: #6b7584;
+  --muted-soft: #8d96a4;
+  --border: #d5dbe4;
+  --border-strong: #b8c2cf;
+  --accent: #2563eb;
+  --accent-strong: #1748b5;
+  --accent-soft: #e7efff;
+  --lime: #b6f13a;
+  --lime-ink: #203300;
+  --stage-bg: #f8fbff;
+  --stage-ink: #07111e;
+  --stage-soft: #33445d;
+  --stage-muted: #6f7b8f;
+  --stage-border: rgba(12, 22, 38, 0.14);
+  --stage-control-bg: rgba(255, 255, 255, 0.58);
+  --stage-control-border: rgba(12, 22, 38, 0.14);
+  --stage-control-hover: rgba(12, 22, 38, 0.06);
+  --stage-action-hover-bg: rgba(12, 22, 38, 0.075);
+  --stage-action-hover-border: rgba(12, 22, 38, 0.22);
+  --stage-action-hover-ink: var(--stage-ink);
+  --stage-primary-action-hover-bg: var(--lime);
+  --stage-primary-action-hover-ink: var(--lime-ink);
+  --stage-glass-bg: rgba(255, 255, 255, 0.54);
+  --stage-glass-bg-hover: rgba(255, 255, 255, 0.7);
+  --stage-selected-bg: #0d1726;
+  --stage-selected-ink: #f7fbff;
+  --appcast-menu-bg: rgba(255, 255, 255, 0.98);
+  --appcast-menu-ink: var(--ink);
+  --appcast-menu-muted: var(--muted);
+  --appcast-menu-border: rgba(12, 22, 38, 0.14);
+  --appcast-menu-hover: rgba(37, 99, 235, 0.08);
+  --appcast-menu-accent: var(--accent-strong);
+  --appcast-menu-shadow: 0 22px 60px rgba(20, 32, 52, 0.16);
+  --code-bg: #e8edf4;
+  --code-border: #c9d2df;
+  --hero-background:
+    linear-gradient(135deg, #f9fbff, #e8f0ff 58%, #f4fbf5);
+  --hero-veil:
+    linear-gradient(90deg, rgba(248, 251, 255, 0.88) 0%, rgba(248, 251, 255, 0.5) 52%, rgba(248, 251, 255, 0.08) 100%);
+  --hero-grid-line: rgba(12, 22, 38, 0.065);
+}
